% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cumulative_competence}
\alias{plot_cumulative_competence}
\title{Cumulative infections against average competence}
\usage{
plot_cumulative_competence(x)
}
\arguments{
\item{x}{A \code{"multihost_sir"} object from \code{\link[=run_multihost_sir]{run_multihost_sir()}}.}
}
\value{
A ggplot object.
}
\description{
Cumulative infections against average competence
}
\examples{
res <- run_multihost_sir(3:7, n_iterations = 10, duration = 150, seed = 1)
plot_cumulative_competence(res)
}
