% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cumulative_richness}
\alias{plot_cumulative_richness}
\title{Cumulative proportion infected by species richness}
\usage{
plot_cumulative_richness(x, max_levels = 8L)
}
\arguments{
\item{x}{A \code{"multihost_sir"} object from \code{\link[=run_multihost_sir]{run_multihost_sir()}}.}

\item{max_levels}{Largest number of richness levels to show.}
}
\value{
A ggplot object.
}
\description{
Cumulative proportion infected by species richness
}
\examples{
res <- run_multihost_sir(3:6, n_iterations = 8, duration = 150, seed = 1)
plot_cumulative_richness(res)
}
