% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_days_to_peak_r0}
\alias{plot_days_to_peak_r0}
\title{Days to peak against community R0}
\usage{
plot_days_to_peak_r0(x)
}
\arguments{
\item{x}{A \code{"multihost_sir"} object from \code{\link[=run_multihost_sir]{run_multihost_sir()}}.}
}
\value{
A ggplot object.
}
\description{
Days to peak against community R0
}
\examples{
res <- run_multihost_sir(3:7, n_iterations = 10, duration = 150, seed = 1)
plot_days_to_peak_r0(res)
}
