% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_peak_density}
\alias{plot_peak_density}
\title{Peak prevalence against total system density}
\usage{
plot_peak_density(x)
}
\arguments{
\item{x}{A \code{"multihost_sir"} object from \code{\link[=run_multihost_sir]{run_multihost_sir()}}.}
}
\value{
A ggplot object.
}
\description{
Peak prevalence against total system density
}
\examples{
res <- run_multihost_sir(3:7, n_iterations = 10, duration = 150, seed = 1)
plot_peak_density(res)
}
