% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_r0_trophic}
\alias{plot_r0_trophic}
\title{Community R0 against average trophic level}
\usage{
plot_r0_trophic(x)
}
\arguments{
\item{x}{A \code{"multihost_sir"} object from \code{\link[=run_multihost_sir]{run_multihost_sir()}}.}
}
\value{
A ggplot object.
}
\description{
Community R0 against average trophic level
}
\examples{
res <- run_multihost_sir(3:7, n_iterations = 10, duration = 150, seed = 1)
plot_r0_trophic(res)
}
